pyproject.toml
[project]
name = "status-validator"
version = "0.1.0"
description = "Automated status validation pipeline using Google Sheets and an LLM"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.118.0",
    "google-auth>=2.28.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.1.0",
    "openai>=1.12.0",
    "pydantic>=2.6.0",
    "python-dateutil>=2.8.2",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock"]

[project.scripts]
status-validator = "status_validator.main:run"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
